correct prototype for C++ compiler target type

+++ bld/plusplus/h/target16.h
@@ -74,5 +74,5 @@ typedef unsigned short  target_uint;
 typedef long            target_long;
 typedef unsigned long   target_ulong;
 
-typedef unsigned long   target_size_t;
+typedef unsigned int    target_size_t;
 typedef unsigned short  target_offset_t;
